netconsd: Fix memory leak at worker thread exit
LeakSanitizer spots the following after netconsd exits:

  Direct leak of 7340032 byte(s) in 2 object(s) allocated from:
    #0 0x7ff6a52a0001 in __interceptor_calloc (/usr/lib/x86_64-linux-gnu/libasan.so.2+0x94001)
    #1 0x40bde2 in create_hashtable /home/calvinow/git/fbkutils/netconsd/worker.c:207
    #2 0x40e2ec in ncrx_worker_thread /home/calvinow/git/fbkutils/netconsd/worker.c:526
    #3 0x7ff6a42e0453 in start_thread (/lib/x86_64-linux-gnu/libpthread.so.0+0x7453)

Obviously, we need to free the memory backing the hashtable in addition
to freeing the struct that holds a pointer to it.

An analogous problem exists in create_hashtable(): that one is actually a
bit embarassing because it matters at runtime. But since realistically we
only resize 4-5 times it isn't terrible.