libncrx: Add separate variables real/monotonic time
Currently, we abuse the ncrx timestamp by passing CLOCK_REALTIME into it,
which is obviously not the right thing to do since libncrx assumes the
timestamp is monotonic.

This introduces a second optional timestamp variable for REALTIME. It
isn't used internally by libncrx at all, it only exists so that an output
pipeline can get a REALTIME timestamp if it likes.

For now, we just update all callers to pass zero to ncrx_process(), a
subsequent patch will teach netconsd how to use this field.