netesto: Initial commit of NEtwork TESting TOols
A framework and set of tools to run distributed network testing. A
particular test consists of multiple instances of netperf running on
"client" machines and sending to "server" machines running netserver.
There is a controller machine which reads a script indicating the
particulars of the test, such as list of clients and servers, types of
transfers (RPC or STREAM), TCP congestion control variants, RPC sizes,
etc. In addition, the controllor copies results and statistics from the
clients and servers and creates graphs. It can also create an html page
with tables and graphs of the results.

Signed-off-by: Lawrence Brakmo <brakmo@fb.com>