netconsd: Fix source of "ncrx: time regressed" messages
I was very pedantic and wrote this to pass the timestamp corresponding to when
the timer callback was schedule to ncrx_process() when running the callbacks.

This is unnecessary, and actually incorrect: ncrx_process() expects that the
timestamp passed to it will not regress, and when it does it prints this:

	ncrx: time regressed: blah->blah

This happens when we get a packet, schedule a callback, get a second packet that
doesn't result in cancelling the callback, and have enough load that we are past
the time the callback was scheduled for when we actually process it.

Fix it by just passing now to ncrx_process() in do_ncrx_callbacks(). While I'm
at it, make the code in run_ncrx_callbacks() a bit clearer.