_no_validate Metrics
Add _no_validate option that if present in a job's metrics
configuration will disable stripping/validation such that whatever a
parser returns is returned directly to the caller of job.run() without
making sure that the correct keys are present and that there are no
extraneous metrics present that were not in the definition.

This is useful for some larger correctness tests like ltp that have a
myriad of different test cases that would be annoying to have to define
all of them in a job config.