Optional subprocess returncode check
Add benchmark option check_returncode (defaults to true) to allow a
parser to still run and return metrics even if the binary exited with a
code other than 0. Correctness tests commonly exit with a code of 1 if
a test fails, but we still want to parse the results and report the
individual failure.