Refactor Benchmark/BenchmarkJob interaction
It was stupid to have a Benchmark class that didn't really do anything,
along with only some configuration options being able to be overridden
by a job configuration. Removed the Benchmark class entirely, just
using a dictionary of benchmark config that is used to setup the
default job configuration, after that a job can override any
configuration option. Renamed 'BenchmarkJob' to just 'Job'

These changes make the api nicer to use outside of the cli context and
clear up potential confusion about which options come from which
configuration.

Additionally, this patch gives Hooks a reference to the Job they are
running for, this allows a Hook to potentially modify any configuration
being used by the Job (for example a Hook might want to restrict the
Job to certain CPUs by changing to the taskset binary)