Fix broken tests
History tests semi-broken due to new hook file format. The new test
makes sure that that the saved file is in a format consistent with the
latest format generated by the code.

Shell hook tests were flaky because they relied on `/tmp` existing and
`/tmp` being the absolute path - this tests fails on MacOS because
`/tmp` is actually a symlink to `/private/tmp`. Use pyfakefs in this
test case to ensure that the results are repeatable cross-platform.