benchpress: simplify metrics handling
Benchpress flattens the test metrics returned from the metrics parser.
E.g.,

{
  'test_a': {
    'success': 1,
    'duration': 0.1,
  },
  'test_b': {
    'success': 0,
    'duration': 0.2,
  },
}

Becomes

{
  'test_a.success': 1,
  'test_a.duration': 0.1,
  'test_b.success': 0,
  'test_b.duration': 0.2,
}

This complicates our internal testing system, which would have to
reverse the flattenning in order to easily report results. Instead,
don't touch the metrics returned by the parser. Reworking the anomaly
detection code is a little too much work, so just pull it out for now;
we can reimplement it to compare the metrics recursively when we need
it.