ncrx: Always terminate msg->text and set msg->dict
Currently, we terminate msg->text and set msg->dict to point at the
dictionary in slot_maybe_complete(). But we don't call it when we end
up retiring OOS messages after the RTO, which means msg->text ends up
containing the entire dictionary:

  scsi 6:0:23:0: Enclosure         facebook Knox2U           0e22 PQ: 0 ANSI: 5\n SUBSYSTEM=scsi\n DEVICE=+scsi:6:0:23:0\n

Fix this by ensuring we always attempt to parse the dictionary before
returning any message to the user.

Signed-off-by: Calvin Owens <calvinowens@fb.com>