[benchpress] add tee_output to job
Summary:
It was very difficult to get the output of benchpress jobs' subprocesses
This adds the capability to add a `tee_output` option to a job config
and have benchpress log the child process's `stdout` and `stderr` to
it's own `stdout`.
Lines from the subprocess are prefixed by either `stdout:` or `stderr:`

The option defaults to False, it can be set to True to log to `stdout`
or to a string indicating the path to a file to write to.

Test Plan:
`python setup.py test`

Reviewers:

Subscribers:

Tasks: T28963009

Tags:

Blame Revision: