netcons: Reassemble LOG_CONT messages properly, kill cont_start
The kernel has a concept of "continuable" messages, as implemented by
the LOG_CONT flag. Previously netcons hasn't had the logic to assemble
these messages, which resulted in missing context for things like
register and state information when getting CPU backtraces, and a whole
host of other lost pieces of context.

This code handles this by introducing a timeout waiting for the
appearance of followup CONT messages (`NCRX_DFL_CONT_MSG_TIMEOUT`,
currently 10000ms) in order to reassemble them to be included into the
current netcons message. We do not include the metadata from these
LOG_CONT messages, we only attempt to merge the textual content included
in the netcons packet. Further sequence numbers are returned with their
text and dict memset to 0 in order to still have the ability to detect
holes in the sequence numbers received.

Here are some questions you may have after reading this patch:

- *Why isn't checking for cont_start enough?* cont_start has also been
  killed as it didn't work properly -- as it was, all LOG_CONT packets
  were improperly being flagged as cont_start, when they should have
  been marked as LOG_CONT. As it is, the kernel doesn't currently have
  any way of flagging these packets.
- *Why do we need `NCRX_DFL_CONT_MSG_TIMEOUT`?* Unfortunately, with the
  current netcons implementation in printk.c, it's actually not possible
  to know which packets will initiate a continuation sequence, which is
  why we wait NCRX_DFL_CONT_MSG_TIMEOUT milliseconds before abandoning
  attempted reassembly.
- *Can't we just fix the kernel?* Yes, and while we can do that, this
  isn't going to fix things for older kernels without changes to
  printk.c, so we still need to be able to handle them.
- *How can a caller know which packets are missing?* The same way we do
  normal sequence skip checking, since other messages are returned as
  husks with their content memset to 0. It is nice (but not required) to
  then skip messages with LOG_CONT set and have `strlen(msg->text) ==
  0` on the caller side.

Test Plan:

1. Set one machine up to forward netcons messages to another using
   bootparams
2. Run ncrx on the other and see `LOG_CONT` messages (in this case from
   `l` to sysrq-trigger) are reassembled before retrieval:

    [18844.440547] RAX: ffffffffffffffda RBX: 0000000000000002 RCX: 00007f5d45db1e7d
    [18844.440548] RDX: 0000000000000000 RSI: 0000000000000000 RDI: 0000000000000004
    [18844.440548] RBP: 0000000000000004 R08: 0000000000000000 R09: 0000000000000050
    [18844.440548] R10: 0000000000000001 R11: 0000000000000293 R12: 00007ffd50c3cdc0
    [18844.440549] R13: 00007ffd50c3d130 R14: 0000000000000000 R15: 0000000000000000
    [18844.440549] Code: 00 00 55 [...]