more explicitly focus on test cases
Summary:
Benchpress started out as a tool to run benchmarks, since then most of the
benchmark-related code has been ripped out.

This diff furthers those changes to move benchpress to a correctness test tool
by creating the TestCaseResult and TestStatus types and moving to that instead
of the ill-defined "metrics" dictionaries thrown around.

The `TestCaseResult` type still allows for passing a `Dict[str, float]` for
`metrics` so that benchmarking functionality can be parsed by benchpress for
use by external tools (or benchpress itself if comparison functionality is
implemented in the future)

Reviewed By: kernelslacker

Differential Revision: D13860231

fbshipit-source-id: 1d216afa7279ab1c07caef62108126b161a0ed61