fbkutils: set details to empty instead of None
Summary:
xfstests have been failing to log results for God knows how long.  Turns out
this is because the testinfra API expects details to be a string, and an empty
string if the test is a success.  We were setting it to None by default which
was throwing an error when we logged the results, meaning we were not getting
results for our xfstests runs.

Reviewed By: vmagro

Differential Revision: D14951510

fbshipit-source-id: 5e1949442c669b4b688b7cabdaa9ef9dad04f463