bsdacctd: Ditch unlink()/O_EXCL logic for opening the acctfile
Summary:
This was a misguided effort to be userproof against using constant paths
in globally-writable directories like /tmp, but as-is it requires us to
force the consumer to restart anytime bsdacctd restarts, which is more
trouble than it's worth.

We don't want O_TRUNC: the consumer's read offset won't have changed, so
our O_APPEND writes will pick up where we left off, and the consumer
will continue happily consuming.

Reviewed By: kernelslacker

Differential Revision: D18377518

fbshipit-source-id: f05b0e782fde7f73597c87607e1f10589de6f822