fbkutils/bsdacctd: Correctly handle acctfile trims after truncation
Summary:
Initially, the acctfile offset and allocated size both start from zero,
and since our target size is a large power of two, this works out: we
only attempt to trim the acctfile when our offset into it is a multiple
of 2MB, which means st_blocks*512 is strictly <= st_size because st_size
is a multiple of the filesystem block size.

Unfortunately, when either bsdacctd restarts or the consumer truncates
the acctfile, this is no longer true: st_size might not be a multiple of
the filesystem block size when we check, and since we subtract st_blocks
from st_size we end up passing the kernel a negative offset:

  fstat(3, {st_blksize=4096, st_blocks=352256, st_size=180352448, ...}) = 0
  fallocate(3, FALLOC_FL_KEEP_SIZE|FALLOC_FL_PUNCH_HOLE, 18446744073709548992, 2097152) = -1 EINVAL (Invalid argument)
  +++ killed by SIGABRT (core dumped) +++

Fix by using an offset of zero instead of doing the math in this case.

Reviewed By: kernelslacker

Differential Revision: D18520319

fbshipit-source-id: c06a99e8ee3e36f1bf23d0f4781fca2221c67793